+++ app/test/shared/user_merger_test.dart
@@ -75,12 +75,14 @@ void main() {
       UserSession()
         ..id = 'target'
         ..userIdKey = hansUser.key
+        ..userId = hansUser.userId
         ..email = 'target@domain.com'
         ..created = DateTime.now()
         ..expires = DateTime.now(),
       UserSession()
         ..id = 'control'
         ..userIdKey = adminUser.key
+        ..userId = adminUser.userId
         ..email = 'control@domain.com'
         ..created = DateTime.now()
         ..expires = DateTime.now(),
@@ -92,8 +94,8 @@ void main() {
       dbService.emptyKey.append(UserSession, id: 'target'),
       dbService.emptyKey.append(UserSession, id: 'control'),
     ]);
-    expect(list[0].userId, joeUser.userId);
-    expect(list[1].userId, adminUser.userId);
+    expect(list[0].userIdValue, joeUser.userId);
+    expect(list[1].userIdValue, adminUser.userId);
   });
 
   testWithServices('new consent', () async {