+++ CHANGELOG.md
@@ -2,6 +2,7 @@ Important changes to data-models, configuration and migrations between each
 AppEngine version, listed here to ease deployment and troubleshooting.
 
 ## Next Release (replace with git tag when deployed)
+ * Deferred: 14 days after the release, remove the `UserSession.userIdKey` property. 
 
 ## `20200910t113209-all`
  * Run `app/bin/tools/check_integrity.dart` **before deploying** the new release.